+++ CHANGES.rst
@@ -1,7 +1,13 @@
+Revision 0.2.0, released XX-07-2018
+-----------------------------------
+
+- Added `dump()` and `dumps()` methods to render Apache configuration
+  back from the `dict`
 
 Revision 0.1.5, released XX-07-2018
 -----------------------------------
 
+- Added config `dump()`/`dumps()` features
 - The `noescape` option implemented
 - Allow empty value syntax in the option-value pair e.g. `option: `
 - Fixed a bug causing malformed AST when a comment resides